# EdnError rework with `parse` rewrite