# Performance
Note this is using the benchmark in `main`. The performance increase should be much better on larger strings, as there is now no `clone()` in parse. The only allocations are for constructing the EDN struct and for parsing numbers. Note that the walker is now all pointer based. Peeking was a big reason previously for the clones, you can see the new performance characteristics (utf-8 complexity but no allocations) https://godbolt.org/z/zK9dEzrYn
## Before
```
     Running benches/parse.rs (target/release/deps/parse-5de3e714a9fe6fe3)
parse                   time:   [23.316 µs 24.270 µs 25.380 µs]
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
```
```
     Running benches/tagged_parse.rs (target/release/deps/tagged_parse-041b63e457dffa70)
tagged_parse            time:   [4.4875 µs 4.5609 µs 4.6395 µs]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
  ```

## After
```
     Running benches/parse.rs (target/release/deps/parse-5de3e714a9fe6fe3)
parse                   time:   [4.6219 µs 4.6649 µs 4.7142 µs]
                        change: [-81.813% -80.967% -80.093%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  8 (8.00%) high severe
  ```
  ```
       Running benches/tagged_parse.rs (target/release/deps/tagged_parse-041b63e457dffa70)
tagged_parse            time:   [1.6429 µs 1.6504 µs 1.6588 µs]
                        change: [-65.166% -64.416% -63.697%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
  ```
# Debug
given
```rust
let edn = "#_ ,, #{hello, this will be discarded} #_{so will this} #{this is invalid";
let e = Edn::from_str(edn);
println!("{e:?}");
```
## Before
`Err(ParseEdn("None could not be parsed at char count 58"))`
## After
`Err(EdnError { code: UnexpectedEOF, line: Some(1), column: Some(74), index: Some(73) })`
Exmples/tests can be found at https://github.com/Grinkers/edn-rs/blob/4365f555f997df08ecffce32a65c18b0dd76870d/tests/error_messages.rs
# Errors
In this rework, the following strings were also previously being parsed as valid EDN. Fuzzing invalid EDN is still a TODO (separate issue/PR).
```clojure
":"
"5011227E71367421E12" ; https://github.com/edn-format/edn#symbols
"{ :[0x42] 42 }"
"\\cats"
"{ :foo 42 :foo 43 }"
```